pyproject.toml
[project]
name = "freee-mcp"
version = "0.1.0"
description = "freee MCP Server - Claude integration with freee accounting API"
requires-python = ">=3.13"
dependencies = [
    "mcp>=0.1.0",
    "requests>=2.31.0",
    "cryptography>=41.0.0",
    "authlib>=1.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
target-version = "py313"
